backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg[binary]==3.2.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic-settings==2.8.1
email-validator==2.2.0
python-multipart==0.0.20
scikit-learn==1.6.1
numpy==2.2.3
joblib==1.4.2
